Improved Bullet Termination
Improved bullet termination by actually terminating when entering an enemy hit box. Due to how the bullet firing system works however in the Ship class, this means that a weapon will get faster the closer a player gets to a target that can have a successful collision. This is explained by the fact that a bullet can only be fired when the previous bullet has terminated its path, so by getting closer to a target, the time till bullet path termination is reduced, therefore increasing fire speed. This can be reworked in later revisions of the game by switching to a bullet objects which can be created and deleted at will in a vector.